+++ NPC.h
@@ -4,20 +4,20 @@
 class NPC {
 private:
     ALLEGRO_BITMAP* npc;
-    int x, y;
+    float x, y;
     int ataca;
     int vida;
     int vidaAct;
     int def;
 public:
     void inicia();
-    void pinta2(int, int);
-    int getx() { return x; };
-    int gety() { return y; };
-    void setx(int _x) { x = _x; };
-    void sety(int _y) { y = _y; };
+    void pinta2(float, float);
+    float getx() { return x; };
+    float gety() { return y; };
+    void setx(float _x) { x = _x; };
+    void sety(float _y) { y = _y; };
     ALLEGRO_BITMAP* getBitmap() { return npc; };
-    void posiciona(int _x, int _y);
+    void posiciona(float _x, float _y);
     bool atacando() { return ataca > 1; };
     void no_ataca() { ataca = -3; };
 };
